+++ pkg/cloudlets/v3/cloudlets.go
@@ -9,12 +9,12 @@ import (
 )
 
 var (
-	// ErrStructValidation is returned when given struct validation failed
+	// ErrStructValidation is returned when given struct validation failed.
 	ErrStructValidation = errors.New("struct validation")
 )
 
 type (
-	// Cloudlets is the api interface for cloudlets
+	// Cloudlets is the api interface for cloudlets.
 	Cloudlets interface {
 		// ListCloudlets returns details of Cloudlets that you can create a shared policy for, including a Cloudlet name and Cloudlet type
 		//
@@ -24,30 +24,37 @@ type (
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/get-policies
 		ListSharedPolicies(context.Context, ListSharedPoliciesRequest) (*ListSharedPoliciesResponse, error)
+
 		// CreateSharedPolicy creates a shared policy for a specific Cloudlet type
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/post-policy
 		CreateSharedPolicy(context.Context, CreateSharedPolicyRequest) (*Policy, error)
+
 		// DeleteSharedPolicy deletes an existing Cloudlets policy
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/delete-policy
 		DeleteSharedPolicy(context.Context, DeleteSharedPolicyRequest) error
+
 		// GetSharedPolicy returns information about a shared policy, including its activation status on the staging and production networks
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/get-policy
 		GetSharedPolicy(context.Context, GetSharedPolicyRequest) (*Policy, error)
+
 		// UpdateSharedPolicy updates an existing policy
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/put-policy
 		UpdateSharedPolicy(context.Context, UpdateSharedPolicyRequest) (*Policy, error)
+
 		// ClonePolicy clones the staging, production, and last modified versions of a non-shared (API v2) or shared policy into a new shared policy
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/post-policy-clone
 		ClonePolicy(context.Context, ClonePolicyRequest) (*Policy, error)
+
 		// ListActivePolicyProperties returns all active properties that are assigned to the policy
 		//
 		// See: https://techdocs.akamai.com/cloudlets/reference/get-policy-properties
 		ListActivePolicyProperties(context.Context, ListActivePolicyPropertiesRequest) (*PolicyProperty, error)
+
 		// ListPolicyVersions lists policy versions by policyID.
 		//
 		// See: https://techdocs.akamai.com/cloudlets/v3/reference/get-policy-versions
@@ -71,21 +78,41 @@ type (
 		// UpdatePolicyVersion updates policy version.
 		//
 		// See: https://techdocs.akamai.com/cloudlets/v3/reference/put-policy-version
+
 		UpdatePolicyVersion(context.Context, UpdatePolicyVersionRequest) (*PolicyVersion, error)
+		// ListPolicyActivations returns the complete activation history for the selected policy in a reverse chronological order.
+		//
+		// See: https://techdocs.akamai.com/cloudlets/reference/get-policy-activations
+		ListPolicyActivations(context.Context, ListPolicyActivationsRequest) (*PolicyActivations, error)
+
+		// ActivatePolicy initiates the activation of the selected cloudlet policy version.
+		//
+		// See: https://techdocs.akamai.com/cloudlets/reference/post-policy-activations
+		ActivatePolicy(context.Context, ActivatePolicyRequest) (*PolicyActivation, error)
+
+		// DeactivatePolicy initiates the deactivation of the selected cloudlet policy version.
+		//
+		// See: https://techdocs.akamai.com/cloudlets/reference/post-policy-activations
+		DeactivatePolicy(context.Context, DeactivatePolicyRequest) (*PolicyActivation, error)
+
+		// GetPolicyActivation activates the selected cloudlet policy version.
+		//
+		// See: https://techdocs.akamai.com/cloudlets/reference/get-policy-activation
+		GetPolicyActivation(context.Context, GetPolicyActivationRequest) (*PolicyActivation, error)
 	}
 
 	cloudlets struct {
 		session.Session
 	}
 
-	// Option defines a Cloudlets option
+	// Option defines a Cloudlets option.
 	Option func(*cloudlets)
 
-	// ClientFunc is a Cloudlets client new method, this can be used for mocking
+	// ClientFunc is a Cloudlets client new method, this can be used for mocking.
 	ClientFunc func(sess session.Session, opts ...Option) Cloudlets
 )
 
-// Client returns a new cloudlets Client instance with the specified controller
+// Client returns a new cloudlets Client instance with the specified controller.
 func Client(sess session.Session, opts ...Option) Cloudlets {
 	c := &cloudlets{
 		Session: sess,